refactor(NES-1659): align dev-hosts lib with project src/lib barrel convention
Other shared libs in this repo (e.g. libs/shared/ai/src/getImageDescription/,
libs/shared/ui/src/libs/extractLanguageNames/) follow the pattern of one
folder per exported feature with its bodies in <feature>.ts plus an
index.ts barrel, and the top-level src/index.ts as a barrel-of-barrels.
The dev-hosts lib previously inlined the function bodies straight into
src/index.ts, which broke that convention.

Move getDevHosts/isDevHost into src/devHosts/devHosts.ts (with its spec
alongside) and re-export through src/devHosts/index.ts and the top-level
src/index.ts barrel. The path alias @core/shared/dev-hosts still resolves
the directory's index.ts, so all consumers (proxy.ts, apolloClient.ts,
getAuthTokens.ts, QrCodeDialog.tsx, both next.config.js files) keep
working unchanged.
