+++ libs/shared/dev-hosts/src/devHosts/index.ts
@@ -1,0 +1,1 @@
+export { getDevHosts, isDevHost } from './devHosts'
